requirements.txt
anthropic
Flask
openai==0.19.0
python-dotenv==0.19.2
requests==2.26.0
